+++ src/account/operations/syncing/mod.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub(crate) mod addresses;
+pub(crate) mod foundries;
 pub mod options;
 pub(crate) mod outputs;
 pub(crate) mod transactions;
@@ -15,7 +16,7 @@ use iota_client::{
     api_types::response::OutputMetadataResponse,
     block::{
         address::{Address, AliasAddress, NftAddress},
-        output::{Output, OutputId},
+        output::{FoundryId, Output, OutputId},
     },
 };
 
@@ -35,7 +36,7 @@ impl AccountHandle {
         log::debug!("[SYNC] start syncing with {:?}", options);
         let syc_start_time = Instant::now();
 
-        // prevent syncing the account multiple times simultaneously
+        // Prevent syncing the account multiple times simultaneously
         let time_now = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .expect("time went backwards")
@@ -47,7 +48,7 @@ impl AccountHandle {
                 "[SYNC] synced within the latest {} ms, only calculating balance",
                 MIN_SYNC_INTERVAL
             );
-            // calculate the balance because if we created a transaction in the meantime, the amount for the inputs is
+            // Calculate the balance because if we created a transaction in the meantime, the amount for the inputs is
             // not available anymore
             return self.balance().await;
         }
@@ -61,7 +62,7 @@ impl AccountHandle {
             Vec<OutputData>,
         ) = self.request_outputs_recursively(addresses_to_sync, &options).await?;
 
-        // request possible spent outputs
+        // Request possible spent outputs
         log::debug!("[SYNC] spent_or_not_synced_outputs: {spent_or_not_synced_output_ids:?}");
         let spent_or_unsynced_output_metadata_responses = self
             .client
@@ -86,7 +87,22 @@ impl AccountHandle {
             self.request_incoming_transaction_data(transaction_ids).await?;
         }
 
-        // updates account with balances, output ids, outputs
+        if options.sync_native_token_foundries {
+            let native_token_foundry_ids = outputs_data
+                .iter()
+                .filter_map(|output| output.output.native_tokens())
+                .flat_map(|native_tokens| {
+                    native_tokens
+                        .iter()
+                        .map(|native_token| FoundryId::from(*native_token.token_id()))
+                })
+                .collect::<HashSet<_>>();
+
+            // Request and store foundry outputs
+            self.request_and_store_foundry_outputs(native_token_foundry_ids).await?;
+        }
+
+        // Updates account with balances, output ids, outputs
         self.update_account(
             addresses_with_unspent_outputs,
             outputs_data,
@@ -101,7 +117,7 @@ impl AccountHandle {
         };
 
         let account_balance = self.balance().await?;
-        // update last_synced mutex
+        // Update last_synced mutex
         let time_now = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .expect("time went backwards")
@@ -127,12 +143,12 @@ impl AccountHandle {
 
         loop {
             let new_outputs_data = if new_alias_and_nft_addresses.is_empty() {
-                // get outputs for addresses and add them also the the addresses_with_unspent_outputs
+                // Get outputs for addresses and add them also the the addresses_with_unspent_outputs
                 let (addresses_with_output_ids, spent_or_not_synced_output_ids_inner) = self
                     .get_output_ids_for_addresses(options, addresses_to_sync.clone())
                     .await?;
                 spent_or_not_synced_output_ids = spent_or_not_synced_output_ids_inner;
-                // get outputs for addresses and add them also the the addresses_with_unspent_outputs
+                // Get outputs for addresses and add them also the the addresses_with_unspent_outputs
                 let (addresses_with_unspent_outputs_inner, outputs_data_inner) = self
                     .get_outputs_from_address_output_ids(addresses_with_output_ids)
                     .await?;