+++ src/message_interface/dtos.rs
@@ -11,7 +11,7 @@ use std::{
 use iota_client::{
     api_types::response::OutputWithMetadataResponse,
     block::{
-        output::OutputId,
+        output::{dto::FoundryOutputDto, FoundryId, OutputId},
         payload::transaction::{dto::TransactionPayloadDto, TransactionId},
     },
 };
@@ -138,6 +138,9 @@ pub struct AccountDto {
     /// Incoming transactions
     #[serde(rename = "incomingTransactions")]
     pub incoming_transactions: HashMap<TransactionId, (TransactionPayloadDto, Vec<OutputWithMetadataResponse>)>,
+    /// Foundries for native tokens in outputs
+    #[serde(rename = "nativeTokenFoundries", default)]
+    pub native_token_foundries: HashMap<FoundryId, FoundryOutputDto>,
 }
 
 impl From<&Account> for AccountDto {
@@ -155,28 +158,30 @@ impl From<&Account> for AccountDto {
                 .collect(),
             outputs: value
                 .outputs()
-                .clone()
-                .into_iter()
-                .map(|(k, o)| (k, OutputDataDto::from(&o)))
+                .iter()
+                .map(|(id, output)| (*id, OutputDataDto::from(output)))
                 .collect(),
             locked_outputs: value.locked_outputs().clone(),
             unspent_outputs: value
                 .unspent_outputs()
-                .clone()
-                .into_iter()
-                .map(|(k, o)| (k, OutputDataDto::from(&o)))
+                .iter()
+                .map(|(id, output)| (*id, OutputDataDto::from(output)))
                 .collect(),
             transactions: value
                 .transactions()
-                .clone()
-                .into_iter()
-                .map(|(k, o)| (k, TransactionDto::from(&o)))
+                .iter()
+                .map(|(id, transaction)| (*id, TransactionDto::from(transaction)))
                 .collect(),
             pending_transactions: value.pending_transactions().clone(),
             incoming_transactions: value
                 .incoming_transactions()
                 .iter()
-                .map(|(tx_id, (tx, inputs))| (*tx_id, (TransactionPayloadDto::from(tx), inputs.clone())))
+                .map(|(id, (transaction, inputs))| (*id, (TransactionPayloadDto::from(transaction), inputs.clone())))
+                .collect(),
+            native_token_foundries: value
+                .native_token_foundries()
+                .iter()
+                .map(|(id, foundry)| (*id, FoundryOutputDto::from(foundry)))
                 .collect(),
         }
     }